refactor in the validations
